Add the ability to set track volumes separately
The way this used to work, track volumes would just be directly applied to note volumes and make everything quiet. With a hint of refactoring and some addControllerEvent later, the program now actually sets the volume of tracks separately. Note that a) I've tested this with LMMS to make sure it would be the same volume but I still don't think `track.volume * 255` (@line 26) isn't the right number and b) it can only go up to 128 so in LMMS terms that's 100% volume